+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v1.0.2
 * Add the ability to put the portal into "Maintenance Mode" by running the
 `toggle_maintenance` Django management command.
 * ehb-datasources upgraded to v1.0.3 -- adds additional Nautilus mappings
+* Added notification of password expiry to LDAP backend
 
 v1.0.1
 ---